Move hpack.h to private
